+++ examples/auth_oauth.rs
@@ -1,12 +1,9 @@
 mod oauth;
 
+use rusty_box::{BoxAPIError, BoxClient, Config, OAuth};
+
+use crate::oauth::{authorize_app, storage};
 use std::env;
-
-use oauth::http_request_listener::request_process;
-use rusty_box::{AuthError, BoxAPIError, BoxClient, Config, OAuth};
-
-static HOSTNAME: &str = "127.0.0.1";
-const PORT: i16 = 5000;
 
 #[tokio::main]
 async fn main() -> Result<(), BoxAPIError> {
@@ -17,32 +14,28 @@ async fn main() -> Result<(), BoxAPIError> {
     let redirect_uri = env::var("REDIRECT_URI").expect("REDIRECT_URI not set");
 
     let config = Config::new();
-    let mut oauth = OAuth::new(
+    let oauth = OAuth::new(
         config,
         client_id,
         client_secret,
-        Some(oauth::storage::save_access_token),
+        Some(storage::save_access_token),
     );
 
-    let (authorization_url, state_out) = oauth.authorization_url(Some(redirect_uri), None, None)?;
+    // Load the OAuth token from the cache file
+    let oauth_json = oauth::storage::load_access_token();
 
-    webbrowser::open(&authorization_url).expect("Failed to open browser");
-
-    let hostname_port = HOSTNAME.to_owned() + ":" + &PORT.to_string();
-    let server = tiny_http::Server::http(hostname_port).unwrap();
-    println!("Listening on {}", server.server_addr());
-
-    let (code, state_in) = match request_process(server) {
-        Ok((code, state)) => (code, state),
-        Err(e) => return Err(BoxAPIError::AuthError(AuthError::Generic(e))),
+    let oauth = match oauth_json {
+        Ok(oauth_json) => {
+            println!("Cached token found, refreshing");
+            let oauth: OAuth =
+                serde_json::from_str(&oauth_json).expect("Failed to parse cached token");
+            oauth
+        }
+        Err(_) => {
+            println!("No cached token found, authorizing app");
+            authorize_app::authorize_app(oauth, Some(redirect_uri)).await?
+        }
     };
-    if state_in != state_out {
-        return Err(BoxAPIError::AuthError(AuthError::Generic(
-            "State mismatch".to_string(),
-        )));
-    }
-
-    oauth.request_access_token(code).await?;
 
     let mut client = BoxClient::new(Box::new(oauth));
