+++ examples/oauth/mod.rs
@@ -1,2 +1,3 @@
+pub mod authorize_app;
 pub mod http_request_listener;
 pub mod storage;
