+++ examples/oauth/storage.rs
@@ -1,7 +1,7 @@
 pub fn save_access_token(json: String) {
-    std::fs::write(".oauth.cache.json", json).expect("Unable to save access token")
+    std::fs::write(".token.cache.json", json).expect("Unable to save access token")
 }
 
-pub fn read_access_token() -> std::io::Result<String> {
-    std::fs::read_to_string(".oauth.cache.json")
+pub fn load_access_token() -> std::io::Result<String> {
+    std::fs::read_to_string(".token.cache.json")
 }
